requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-filter==22.1
django-on-heroku==1.1.2
django-online-users==0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
